Update pythonfunction.yml.template runner
The runner ubuntu-18 is too old and I wasn't able to run it while going through the lab as GH is never able to find a runner that meets the requirements. I had to updated to latest and it worked fine. Sharing it here so that no one else has the same issue.